Change double to float for improved performance and memory efficiency in visualize.c and net.h

+++ net.h
@@ -11,43 +11,43 @@
 
 typedef struct {
     // Weight matrices
-    double W1[HIDDEN_DIM][INPUT_DIM];
-    double W2[OUTPUT_DIM][HIDDEN_DIM];
+    float W1[HIDDEN_DIM][INPUT_DIM];
+    float W2[OUTPUT_DIM][HIDDEN_DIM];
     
     // Layer activations
-    double h[3][HIDDEN_DIM];  // h[0] is input, h[1] is hidden, h[2] is output
+    float h[3][HIDDEN_DIM];  // h[0] is input, h[1] is hidden, h[2] is output
     
     // Gradient accumulation
-    double dW1[HIDDEN_DIM][INPUT_DIM];
-    double dW2[OUTPUT_DIM][HIDDEN_DIM];
+    float dW1[HIDDEN_DIM][INPUT_DIM];
+    float dW2[OUTPUT_DIM][HIDDEN_DIM];
     
     // Sophia parameters
-    double m1[HIDDEN_DIM][INPUT_DIM];  // Momentum
-    double m2[OUTPUT_DIM][HIDDEN_DIM];
-    double h1[HIDDEN_DIM][INPUT_DIM];  // Hessian estimates
-    double h2[OUTPUT_DIM][HIDDEN_DIM];
-    double prev_dW1[HIDDEN_DIM][INPUT_DIM];  // Previous gradients
-    double prev_dW2[OUTPUT_DIM][HIDDEN_DIM];
+    float m1[HIDDEN_DIM][INPUT_DIM];  // Momentum
+    float m2[OUTPUT_DIM][HIDDEN_DIM];
+    float h1[HIDDEN_DIM][INPUT_DIM];  // Hessian estimates
+    float h2[OUTPUT_DIM][HIDDEN_DIM];
+    float prev_dW1[HIDDEN_DIM][INPUT_DIM];  // Previous gradients
+    float prev_dW2[OUTPUT_DIM][HIDDEN_DIM];
     unsigned long long t;
     
-    double lr;
-    double beta1;    // Momentum decay
-    double beta2;    // Hessian decay
-    double rho;      // Hessian clipping
-    double epsilon;
+    float lr;
+    float beta1;    // Momentum decay
+    float beta2;    // Hessian decay
+    float rho;      // Hessian clipping
+    float epsilon;
 } Net;
 
-static double gelu(double x) {
+static float gelu(float x) {
     return 0.5 * x * (1 + tanh(sqrt(2/M_PI) * (x + 0.044715 * pow(x, 3))));
 }
 
-static double gelu_derivative(double x) {
-    double cdf = 0.5 * (1 + tanh(sqrt(2/M_PI) * (x + 0.044715 * pow(x, 3))));
-    double pdf = exp(-0.5 * x * x) / sqrt(2 * M_PI);
+static float gelu_derivative(float x) {
+    float cdf = 0.5 * (1 + tanh(sqrt(2/M_PI) * (x + 0.044715 * pow(x, 3))));
+    float pdf = exp(-0.5 * x * x) / sqrt(2 * M_PI);
     return cdf + x * pdf;
 }
 
-Net* create_net(double learning_rate) {
+Net* create_net(float learning_rate) {
     Net* net = (Net*)calloc(1, sizeof(Net));
     if (!net) return NULL;
 
@@ -59,30 +59,30 @@ Net* create_net(double learning_rate) {
     net->t = 0;
 
     // Xavier initialization
-    double scale1 = sqrt(2.0 / (INPUT_DIM + HIDDEN_DIM));
-    double scale2 = sqrt(2.0 / (HIDDEN_DIM + OUTPUT_DIM));
+    float scale1 = sqrt(2.0 / (INPUT_DIM + HIDDEN_DIM));
+    float scale2 = sqrt(2.0 / (HIDDEN_DIM + OUTPUT_DIM));
 
     for (int i = 0; i < HIDDEN_DIM; i++) {
         for (int j = 0; j < INPUT_DIM; j++) {
-            net->W1[i][j] = ((double)rand()/RAND_MAX * 2 - 1) * scale1;
+            net->W1[i][j] = ((float)rand()/RAND_MAX * 2 - 1) * scale1;
         }
     }
 
     for (int i = 0; i < OUTPUT_DIM; i++) {
         for (int j = 0; j < HIDDEN_DIM; j++) {
-            net->W2[i][j] = ((double)rand()/RAND_MAX * 2 - 1) * scale2;
+            net->W2[i][j] = ((float)rand()/RAND_MAX * 2 - 1) * scale2;
         }
     }
 
     return net;
 }
 
-void forward_net(Net* net, const double* input) {
+void forward_net(Net* net, const float* input) {
     // Copy input
-    memcpy(net->h[0], input, INPUT_DIM * sizeof(double));
+    memcpy(net->h[0], input, INPUT_DIM * sizeof(float));
 
     // Hidden layer
-    memset(net->h[1], 0, HIDDEN_DIM * sizeof(double));
+    memset(net->h[1], 0, HIDDEN_DIM * sizeof(float));
     for (int i = 0; i < HIDDEN_DIM; i++) {
         for (int j = 0; j < INPUT_DIM; j++) {
             net->h[1][i] += net->W1[i][j] * net->h[0][j];
@@ -91,7 +91,7 @@ void forward_net(Net* net, const double* input) {
     }
 
     // Output layer
-    memset(net->h[2], 0, OUTPUT_DIM * sizeof(double));
+    memset(net->h[2], 0, OUTPUT_DIM * sizeof(float));
     for (int i = 0; i < OUTPUT_DIM; i++) {
         for (int j = 0; j < HIDDEN_DIM; j++) {
             net->h[2][i] += net->W2[i][j] * net->h[1][j];
@@ -99,8 +99,8 @@ void forward_net(Net* net, const double* input) {
     }
 }
 
-void backward_net(Net* net, const double* output_gradients) {
-    double delta[HIDDEN_DIM];
+void backward_net(Net* net, const float* output_gradients) {
+    float delta[HIDDEN_DIM];
     
     // Output layer gradients
     for (int i = 0; i < OUTPUT_DIM; i++) {
@@ -110,7 +110,7 @@ void backward_net(Net* net, const double* output_gradients) {
     }
 
     // Hidden layer gradients
-    memset(delta, 0, HIDDEN_DIM * sizeof(double));
+    memset(delta, 0, HIDDEN_DIM * sizeof(float));
     for (int i = 0; i < HIDDEN_DIM; i++) {
         for (int j = 0; j < OUTPUT_DIM; j++) {
             delta[i] += output_gradients[j] * net->W2[j][i];
@@ -138,8 +138,8 @@ void update_net(Net* net) {
             net->m1[i][j] = net->beta1 * net->m1[i][j] + (1.0 - net->beta1) * net->dW1[i][j];
             
             // Approximate second derivative using consecutive gradients
-            double grad_diff = net->dW1[i][j] - net->prev_dW1[i][j];
-            double hessian_approx = fabs(grad_diff / (net->lr + net->epsilon));
+            float grad_diff = net->dW1[i][j] - net->prev_dW1[i][j];
+            float hessian_approx = fabs(grad_diff / (net->lr + net->epsilon));
             
             // Update and clip Hessian estimate
             net->h1[i][j] = fmax(net->rho, 
@@ -161,8 +161,8 @@ void update_net(Net* net) {
             net->m2[i][j] = net->beta1 * net->m2[i][j] + (1.0 - net->beta1) * net->dW2[i][j];
             
             // Approximate second derivative using consecutive gradients
-            double grad_diff = net->dW2[i][j] - net->prev_dW2[i][j];
-            double hessian_approx = fabs(grad_diff / (net->lr + net->epsilon));
+            float grad_diff = net->dW2[i][j] - net->prev_dW2[i][j];
+            float hessian_approx = fabs(grad_diff / (net->lr + net->epsilon));
             
             // Update and clip Hessian estimate
             net->h2[i][j] = fmax(net->rho, 
@@ -183,11 +183,11 @@ bool save_net(const char* filename, const Net* net) {
     if (!file) return false;
     
     // Save parameters
-    fwrite(&net->lr, sizeof(double), 1, file);
-    fwrite(&net->beta1, sizeof(double), 1, file);
-    fwrite(&net->beta2, sizeof(double), 1, file);
-    fwrite(&net->rho, sizeof(double), 1, file);
-    fwrite(&net->epsilon, sizeof(double), 1, file);
+    fwrite(&net->lr, sizeof(float), 1, file);
+    fwrite(&net->beta1, sizeof(float), 1, file);
+    fwrite(&net->beta2, sizeof(float), 1, file);
+    fwrite(&net->rho, sizeof(float), 1, file);
+    fwrite(&net->epsilon, sizeof(float), 1, file);
     fwrite(&net->t, sizeof(unsigned long long), 1, file);
     
     // Save weights and Sophia states
@@ -215,11 +215,11 @@ Net* load_net(const char* filename) {
     }
     
     // Load parameters
-    if (fread(&net->lr, sizeof(double), 1, file) != 1 ||
-        fread(&net->beta1, sizeof(double), 1, file) != 1 ||
-        fread(&net->beta2, sizeof(double), 1, file) != 1 ||
-        fread(&net->rho, sizeof(double), 1, file) != 1 ||
-        fread(&net->epsilon, sizeof(double), 1, file) != 1 ||
+    if (fread(&net->lr, sizeof(float), 1, file) != 1 ||
+        fread(&net->beta1, sizeof(float), 1, file) != 1 ||
+        fread(&net->beta2, sizeof(float), 1, file) != 1 ||
+        fread(&net->rho, sizeof(float), 1, file) != 1 ||
+        fread(&net->epsilon, sizeof(float), 1, file) != 1 ||
         fread(&net->t, sizeof(unsigned long long), 1, file) != 1 ||
         
         // Load weights and Sophia states
